+++ src/Domain.LinnApps/WwdWorkDetail.cs
@@ -8,7 +8,7 @@
 
         public string LocationGroup { get; set; }
 
-        public decimal? Quantity { get; set; }
+        public int? Quantity { get; set; }
 
         public string State { get; set; }
     }
